# arion/docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateOrderRequest:
    properties:
      customer_id:
        type: integer
      notes:
        type: string
    type: object
...
  handler.ErrorApiResponse:
    properties:
      code:
        type: string
      data:
        type: object
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handler.UpdateOrderItemRequest:
    properties:
      product_id:
        type: integer
      qty:
        type: integer
    type: object
  handler.UpdateOrderRequest:
    properties:
      customer_id:
        type: integer
      notes:
        type: string
      status:
        type: string
      total_price:
        type: integer
    type: object
...
  response.OrderData:
    properties:
      created_at:
        type: string
      customer_name:
        type: string
      id:
        type: integer
      notes:
        type: string
      order_items:
        items:
          $ref: '#/definitions/response.OrderItemData'
        type: array
      status:
        type: string
      total_price:
        type: integer
      updated_at:
        type: string
    type: object
  response.OrderItemData:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      price:
        type: integer
      product_name:
        type: string
      qty:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: |-
    Order management API for Jastipers

    All responses use this envelope:
    { "success": bool, "data": <payload>, "code": string, "message": string }
    Success: success=true, data holds the payload. Error: success=false, data={}, code and message describe the error.
  title: Recapo API
  version: "1.0"
paths:
...
  /order:
    post:
      consumes:
      - application/json
      description: |-
        Create a new order for the shop.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Order data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderData'
        "400":
          description: Bad request (invalid JSON or validation)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - order
  /orders:
    get:
      consumes:
      - application/json
      description: |-
        Get all orders for the shop. Optional search query to filter.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.OrderData'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - order
  /orders/{order_id}:
    delete:
      consumes:
      - application/json
      description: |-
        Delete an order by ID.
        Success Response envelope: { success, data, code, message }. data contains "OK" on success.
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success. data contains \"OK\
          schema:
            type: string
        "400":
          description: Bad request (invalid order_id)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Delete order
      tags:
      - order
    get:
      consumes:
      - application/json
      description: |-
        Get a single order by ID.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderData'
        "400":
          description: Bad request (invalid order_id)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - order
    patch:
      consumes:
      - application/json
      description: |-
        Update an existing order. Only provided fields are updated.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderData'
        "400":
          description: Bad request (invalid JSON or order_id)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Update order
      tags:
      - order
  /orders/{order_id}/items:
    get:
      consumes:
      - application/json
      description: |-
        Get all items for an order.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.OrderItemData'
            type: array
        "400":
          description: Bad request (invalid order_id)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: List order items
      tags:
      - order
  /orders/{order_id}/items/{item_id}:
    delete:
      consumes:
      - application/json
      description: |-
        Delete an order item by ID.
        Success Response envelope: { success, data, code, message }. data contains "OK" on success.
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: Order item ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success. data contains \"OK\
          schema:
            type: string
        "400":
          description: Bad request (invalid order_id or item_id)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Delete order item
      tags:
      - order
    get:
      consumes:
      - application/json
      description: |-
        Get a single order item by ID.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: Order item ID
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderItemData'
        "400":
          description: Bad request (invalid order_id or item_id)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Get order item by ID
      tags:
      - order
    patch:
      consumes:
      - application/json
      description: |-
        Update an order item. Only provided fields are updated.
        Success Response envelope: { success, data, code, message }. Schema below shows the data field (inner payload).
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: integer
      - description: Order item ID
        in: path
        name: item_id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateOrderItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderItemData'
        "400":
          description: Bad request (invalid JSON, order_id, or item_id)
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorApiResponse'
      security:
      - BearerAuth: []
      summary: Update order item
      tags:
      - order
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"